Add files via upload
Census statistics are collected every ten years to examine the impact of population
expansion on various characteristics. Understanding the differences in life expectancy
between continents can also help us better understand how lifestyle, education, and
access to high-quality healthcare services have affected it throughout the years. It’s been
really interesting to realize the difference, where we are 20 years from now, especially
with the global fight against the dreaded COVID-19.
This research examines 227 nations to identify regional variations or convergences in
life expectancy at birth and under-five mortality rates between 2002 and 2022. To this
end, we first performed a uni-variate analysis of all numeric variables in 2022. Bi-variate
associations between these numerical variables are also visible. We also investigate the
relationships between and within the various sub-regions of these variables.
The project report that is being presented here has 4 additional sections in addition to
the introductory section. The description of the used data set, including the definitions
of the variables, and the calibre of the data provided, is covered in Section 2. The
definition of various statistical methods needed for data analysis, such as mean, median,
variance, etc., is covered in Section 3. In addition to Correlation, here is an explanation
of the concept of correlation. Interpreting the results of the final analysis using the
statistical methods discussed in Section 3 is covered in Section 4, which follows. The
final section: Section 5 serves as the summation. It includes an overview of all the results
and explanations. It also offers a perspective for a more in-depth investigation.